# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kosova Estate",
        "version": "1.0.0"
    },
    "paths": {
        "/api/agents": {
            "get": {
                "summary": "Get all agents",
                "operationId": "defdc8451e3eedb011a818c7dd227d6f",
                "responses": {
                    "200": {
                        "description": "List of agents"
                    },
                    "500": {
                        "description": "Unable to fetch agents"
                    }
                }
            }
        },
        "/api/agents/{id}": {
            "put": {
                "summary": "Edit an agent",
                "operationId": "7d003d53937ed09545740a185148a5da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Agent ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "First name of the agent",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Last name of the agent",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email of the agent",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Phone number of the agent",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Username of the agent",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "job_position",
                        "in": "query",
                        "description": "Job position of the agent",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "Image path of the agent",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "500": {
                        "description": "Unable to update user"
                    }
                }
            },
            "delete": {
                "summary": "Delete an agent",
                "operationId": "c144bff0e057fe112adca1005b658e7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Agent ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    },
                    "500": {
                        "description": "Unable to delete user"
                    }
                }
            }
        },
        "/api/agents/upload-image": {
            "post": {
                "summary": "Upload an agent image",
                "operationId": "7c03a81cd0a52b9150fa6d5dec3a98be",
                "requestBody": {
                    "description": "Image file to upload",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "Image file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Image uploaded successfully"
                    },
                    "400": {
                        "description": "No file uploaded"
                    },
                    "500": {
                        "description": "Unable to upload image"
                    }
                }
            }
        },
...
    }
}